% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{bin}
\alias{bin}
\title{Bin numeric columns.}
\usage{
bin(vars = numerics, breaks = 10)
}
\arguments{
\item{vars}{Function or formula that returns selected columns from a data.frame. Alternatively, character vector of column names.}

\item{breaks}{Either an integer representing the number of bins or a vector of break points.}
}
\description{
Bin numeric columns.
}
\examples{
df <- data.frame(A=rep("a", 100), B=rnorm(100))
prep <- bin()
prep$fit(df)
prep$transform(df)
# or more succintly
prep$fit_transform(df)
}
