% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{characters}
\alias{characters}
\title{Helper function to select character variables.}
\usage{
characters(data)
}
\description{
Helper function to select character variables.
}
\examples{
df <- data.frame(a=c(1,2,NA,3), b=c(10,20,30,NA), c=c("c1", NA, "c3", "c3"))
prep <- impute(characters, most_frequent)
prep$fit(df)
prep$transform(df)
}

