% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{create_histograms}
\alias{create_histograms}
\title{Create Histograms for Numeric Columns}
\usage{
create_histograms(data, bins = 30, facet = FALSE)
}
\arguments{
\item{data}{A dataframe.}

\item{bins}{Number of bins for the histogram.}

\item{facet}{Boolean; if TRUE, create a single faceted plot for all histograms.}
}
\description{
This function creates histograms for all numeric columns in a given dataframe.
Optionally, it can create a single plot with a histogram for each numeric column using facets.
}
\examples{
create_histograms(mtcars)
create_histograms(iris, bins = 20)
create_histograms(mtcars, facet = TRUE)
}
