% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_tools.R
\name{createBarCharts}
\alias{createBarCharts}
\title{Create Bar Charts for Categorical Columns}
\usage{
createBarCharts(data, facet = FALSE)
}
\arguments{
\item{data}{A dataframe containing the data to be plotted.}

\item{facet}{Logical; if TRUE, create a single faceted plot for all bar charts.
If FALSE, create individual plots for each categorical column. Defaults to FALSE.}
}
\value{
An invisible list of ggplot objects if `facet` is FALSE. If `facet` is TRUE,
        a single ggplot object is returned. Each plot represents the frequency
        distribution of a categorical variable in the dataframe.
}
\description{
This function creates bar charts for all categorical columns in a given dataframe.
Optionally, it can create a single plot with a bar chart for each categorical column using facets.
}
\examples{
createBarCharts(mtcars, facet = FALSE)
createBarCharts(iris, facet = TRUE)

}
