% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_tools.R
\name{createBoxplot}
\alias{createBoxplot}
\title{Create Boxplot for a Specified Column}
\usage{
createBoxplot(data, column)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{column}{The name of the column to plot.}
}
\value{
A ggplot object representing the boxplot of the column.
}
\description{
This function creates a boxplot for a specified column in a data frame,
which is useful for visualizing the distribution and identifying outliers.
}
\examples{
createBoxplot(mtcars, "mpg")
}
