% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_tools.R
\name{createHistograms}
\alias{createHistograms}
\title{Create Histograms for Numeric Columns}
\usage{
createHistograms(data, bins = 30)
}
\arguments{
\item{data}{A dataframe.}

\item{bins}{Number of bins for the histogram.}
}
\description{
This function creates histograms for all numeric columns in a given dataframe.
}
\examples{
createHistograms(mtcars)
createHistograms(iris, bins = 20)
}
