% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{encode_impact}
\alias{encode_impact}
\title{Create impact encoder, that turns factor variables into integers.
Where as the numeric encoder attributes arbitrary integers to the factor levels,
the impact encoder produces ordered levels (with respect to the target).}
\usage{
encode_impact(vars = categoricals, target)
}
\arguments{
\item{vars}{Function or formula that returns selected columns from a data.frame. Alternatively, character vector of column names.}

\item{target}{Name of the target variable.}
}
\description{
Create impact encoder, that turns factor variables into integers.
Where as the numeric encoder attributes arbitrary integers to the factor levels,
the impact encoder produces ordered levels (with respect to the target).
}
\details{
An approximate description of the algorithm is that the levels of the factors are ordered according to the mean value of the target for each level,
but credibility is also taken into account. Under the hood, a model with random effects is used.
}
\examples{
df <- data.frame(A=sample(c('a1', 'a2'), 100, replace=TRUE), y = runif(100))
prep <- encode_impact(vars = 'A', target = 'y')
prep$fit(df)
prep$transform(df)
# or more succintly
prep$fit_transform(df)
}
