% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{encode_numeric}
\alias{encode_numeric}
\title{Create numeric encoder, that turns factor variables into integers.}
\usage{
encode_numeric(vars = factors)
}
\arguments{
\item{vars}{Function or formula that returns selected columns from a data.frame. Alternatively, character vector of column names.}
}
\description{
Create numeric encoder, that turns factor variables into integers.
}
\examples{
df <- data.frame(A=c("a1","a1","a2","a3"), B=c("b1", "b2", "b3", "b4"), X=c(1,2,3,4))
prep <- encode_numeric(~select(., A))
prep$fit(df)
prep$transform(df)
# or more succintly
prep$fit_transform(df)
}
