% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_tools.R
\name{estimateBoxCoxLambda}
\alias{estimateBoxCoxLambda}
\title{Estimate Optimal Lambda for Box-Cox Transformation}
\usage{
estimateBoxCoxLambda(data, column_name)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{column_name}{The name of the column for which to estimate lambda.}
}
\value{
The estimated lambda value for the Box-Cox transformation.
}
\description{
This function estimates the optimal lambda parameter for the Box-Cox transformation
of a specified column in a dataset. It requires all values in the column to be positive.
}
\examples{
data(iris)
lambda_value <- estimateBoxCoxLambda(iris, "Sepal.Length")
print(lambda_value)
}
