% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{factorize}
\alias{factorize}
\title{Make harmonized factor variables.}
\usage{
factorize(vars = categoricals)
}
\arguments{
\item{vars}{Function or formula that returns selected columns from a data.frame. Alternatively, character vector of column names.}
}
\description{
This function makes selected variables as factors and also harmonizes the levels of a new dataset to match those of the fitted dataset.
This is important when a factor variable in a test set contains levels that were not present for that variable during training.
In that case, the new levels are replaced by the most frequent value found during training.
}
\examples{
df <- data.frame(A=c("a1","a1","a1","a2"), stringsAsFactors = FALSE)
df2 <- data.frame(A=c("a1","a1","a4","a5"), stringsAsFactors = FALSE)
prep <- factorize()
prep$fit(df)
prep$transform(df2)
}

