% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{factors}
\alias{factors}
\title{Helper function to select factor variables.}
\usage{
factors(data)
}
\description{
Helper function to select factor variables.
}
\examples{
df <- data.frame(a=c(1,2,NA,3), b=c(10,20,30,NA), c=c("c1", NA, "c3", "c3"))
prep <- impute(factors, most_frequent)
prep$fit(df)
prep$transform(df)
}
