% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{fuse_one_hot}
\alias{fuse_one_hot}
\title{Take a set of one hot encoded columns (values: 0 or 1) and transform them back into a single factor column.}
\usage{
fuse_one_hot(vars, out = ".out", missing = "missing", keep = FALSE)
}
\arguments{
\item{vars}{Function or formula that returns selected columns from a data.frame. Alternatively, character vector of column names.}

\item{out}{Name of the new column}

\item{missing}{Value to use for the rows where all the one hot columns are zero.}

\item{keep}{Should the original one hot encoded column be kept}
}
\description{
Take a set of one hot encoded columns (values: 0 or 1) and transform them back into a single factor column.
}
\examples{
df <- data.frame(A=c("a1","a1","a2","a3"), B=c(1, 0, 0, 1), C=c(0, 1, 0, 0))
prep <- fuse_one_hot(vars = c('B', 'C'))
prep$fit(df)
prep$transform(df)
# or more succintly
prep$fit_transform(df)
}

