% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{map_values}
\alias{map_values}
\title{Map values of variables.}
\usage{
map_values(vars, mapping, rename)
}
\arguments{
\item{vars}{Function or formula that returns selected columns from a data.frame. Alternatively, character vector of column names.}

\item{mapping}{The mapping to apply. This should be a named vector, where the names are the old values, and the elements are the new values.
There should be at most one unnamed element, representing `all other` elements.}

\item{func}{Function to do the imputation (e.g. mean, median).
Can be a user defined function or a formula, where a dot represents the column.}
}
\description{
Map values of variables.
}
\details{
Currently, this function always turns the mapped variables into factors.
}
\examples{
df <- data.frame(a=c(1,2,NA,3), b=c(10,20,30,NA), c=c("c1", "c2", "c3", "c4"))
prep <- map_values("c", c(c1="C1", c4="C4", "other"))
prep$fit(df)
prep$transform(df)
# or more succintly
prep$fit_transform(df)
}

