% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_tools.R
\name{missing_data_proportion}
\alias{missing_data_proportion}
\title{Calculate Proportion of Missing Data}
\usage{
missing_data_proportion(data)
}
\arguments{
\item{data}{A dataframe representing the dataset.}
}
\value{
A named vector with predictors as names and proportions of missing data as values.
}
\description{
This function calculates the proportion of missing data for each predictor in the dataset.
It returns a named vector where names are the predictors and values are the proportions
of missing data.
}
\examples{
data("Soybean", package = "mlbench")
missing_proportions <- missing_data_proportion(Soybean)
print(missing_proportions)
}
