% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{most_frequent}
\alias{most_frequent}
\title{Helper function to find the most frequent values in a variable.}
\usage{
most_frequent(v, n = 1)
}
\arguments{
\item{v}{Variable for which you want the top-n most frequent values.}

\item{n}{How many values to return.}
}
\description{
Helper function to find the most frequent values in a variable.
}
