% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{numerics}
\alias{numerics}
\title{Helper function to select numeric variables.}
\usage{
numerics(data)
}
\description{
Helper function to select numeric variables.
}
\examples{
df <- data.frame(a=c(1,2,NA,3), b=c(10,20,30,NA), c=c("c1", "c2", "c3", "c4"))
prep <- impute(numerics, mean, na.rm = TRUE)
prep$fit(df)
prep$transform(df)
}
