% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_tools.R
\name{plot_IQR}
\alias{plot_IQR}
\title{Plot IQR and Outliers}
\usage{
plot_IQR(data, column)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{column}{The name of the column to plot.}
}
\value{
A ggplot object representing the boxplot of the column.
}
\description{
This function creates a boxplot for a specified column in a data frame,
highlighting the IQR and outliers.
}
\examples{
# example code
plot_IQR(mtcars, "mpg")
}
