\name{preprocessr-package}
\alias{preprocessr-package}
\alias{preprocessr}
\docType{package}
\title{
\packageTitle{preprocessr}
}
\description{
\packageDescription{preprocessr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{preprocessr}
\packageIndices{preprocessr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{preprocessr}

Maintainer: \packageMaintainer{preprocessr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
