% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{remove_constants}
\alias{remove_constants}
\title{Remove constant columns.}
\usage{
remove_constants(vars = allvars)
}
\arguments{
\item{vars}{Function or formula that returns selected columns from a data.frame. Alternatively, character vector of column names.}
}
\description{
Remove constant columns.
}
\examples{
df <- data.frame(A=c("a1","a1","a2","a3"), B=c("b1", "b2", "b3", "b4"), C=c("c","c","c","c"), X=c(1,1,1,1))
prep <- remove_constants()
prep$fit(df)
prep$transform(df)
# or more succintly
prep$fit_transform(df)
}
