% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{replace_values}
\alias{replace_values}
\title{Replace specific values in a column (usually NAs).}
\usage{
replace_values(vars = numerics, values = NA, func, ...,
  rename = function(x) x)
}
\arguments{
\item{vars}{Function or formula that returns selected columns from a data.frame. Alternatively, character vector of column names.}

\item{values}{Vector of values to replace. Default is NA.}

\item{func}{Function to do the replacement (e.g. mean, median).
Can be a constant, a user defined function or a formula, where a dot represents the column.}

\item{...}{Other arguments passed to \code{func}.}

\item{rename}{Function to rename the output variables (input: the original column name, output: the new column name).}
}
\description{
Replace specific values in a column (usually NAs).
}
\examples{
df <- data.frame(a=c(1,2,NA,4), b=c(10,20,40,NA), c=c("c1", "c2", "c3", "c4"))
prep <- replace_values(numerics, c(NA, 0), mean, na.rm = TRUE, rename = ~ paste(., "interim", sep="_"))
prep$fit(df)
prep$transform(df)
# or more succintly
prep$fit_transform(df)
}
