% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.R
\name{scale_minmax}
\alias{scale_minmax}
\title{Scale columns so that values lie between min and max value.}
\usage{
scale_minmax(vars = numerics)
}
\arguments{
\item{vars}{Function or formula that returns selected columns from a data.frame. Alternatively, character vector of column names.}
}
\description{
Scale columns so that values lie between min and max value.
}
\examples{
df <- data.frame(A=c("a1","a1","a2","a3"), B=c("b1", "b2", "b3", "b4"), X=c(1,2,3,4))
prep <- scale()
prep$fit(df)
prep$transform(df)
# or more succintly
prep$fit_transform(df)
}

