% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessr.R
\name{soundex}
\alias{soundex}
\title{Get soundex code for a given string}
\usage{
soundex(string, ...)
}
\description{
Since soundex is only defined on the letters a-z, this function will first
convert any input to ASCII before calculating the soundex.
}

