% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessr.R
\name{tidy_string}
\alias{tidy_string}
\title{Truncate a string to a standard size}
\usage{
tidy_string(string, length = -1, pattern = "[^A-Za-z0-9]",
  lowercase = TRUE)
}
\arguments{
\item{string}{character vector of strings to truncate}

\item{length}{how long (at most) the resulting strings should be. -1
(default) to not truncate.}

\item{pattern}{regex pattern for characters that should be removed.}
}
\description{
It's often useful in doing partial string comparisons (eg, for comparing
street addresses) to remove non-standard characters and reduce the length.
}
\details{
By default, only alpha-numeric characters are retained
}
\examples{
address1 <- "123 N Washington way"
address2 <- "123 N. Washington"
identical(str_truncate(address1, 8), str_truncate(address2, 8))
}

