% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_tools.R
\name{calculate_IQR}
\alias{calculate_IQR}
\title{Calculate Interquartile Range (IQR) and Identify Outliers}
\usage{
calculate_IQR(data, column)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{column}{The name of the column for which to calculate the IQR.}
}
\value{
A list with six elements: Column (the name of the column),
        Q1 (first quartile), Q3 (third quartile), IQR (the interquartile range),
        Lower Bound, Upper Bound, and Outliers (vector of outlier values).
}
\description{
This function calculates the interquartile range (IQR) for a specified column
in a data frame and identifies any outliers. It also returns the first and third
quartiles, and the lower and upper bounds used for outlier detection.
}
\examples{
calculate_IQR(mtcars, "mpg")

# To apply calculate_IQR to all numeric columns in a dataframe:
numeric_columns <- sapply(mtcars, is.numeric)
IQR_results <- lapply(names(mtcars)[numeric_columns], function(col) calculate_IQR(mtcars, col))

}
