% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{create_boxplots}
\alias{create_boxplots}
\title{Create Boxplots for Numeric Columns}
\usage{
create_boxplots(data, facet = FALSE)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{facet}{Logical; if TRUE, create a single faceted plot for all boxplots.}
}
\value{
A list of ggplot objects representing the boxplots of the numeric columns, or a single ggplot object if faceting.
}
\description{
This function creates a boxplot for each numeric column in a data frame.
Optionally, it can create a single plot with a boxplot for each numeric column using facets.
}
\examples{
createBoxplots(mtcars)
createBoxplots(mtcars, facet = TRUE)
}
