% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_tools.R
\name{createCorrPlot}
\alias{createCorrPlot}
\title{Create Correlation Plot}
\usage{
createCorrPlot(data)
}
\arguments{
\item{data}{A dataframe containing the data.}
}
\description{
This funciton creates a correlation plot for all numeric columns in a given dataframe
Non-numeric columns are automatically excluded.
}
\examples{
data(mtcars)
createCorrPlot(mtcars)
}
